+++ crates/module-system/module-implementations/sov-bank/tests/integration/capability_test.rs
@@ -2,7 +2,7 @@ use sov_bank::{config_gas_token_id, Bank, Coins, ReserveGasError};
 use sov_modules_api::prelude::UnwrapInfallible;
 use sov_modules_api::transaction::PriorityFeeBips;
 use sov_modules_api::{Gas, GasUnit, Spec, TxEffect};
-use sov_test_utils::{get_gas_used, AsUser, SkippedReason, TestUser, TransactionTestCase};
+use sov_test_utils::{get_gas_used, AsUser, TestUser, TransactionTestCase, TxProcessingError};
 
 use crate::helpers::{setup, TestData};
 
@@ -236,7 +236,9 @@ fn test_reserve_gas_no_account() {
             },
         }),
         assert: Box::new(move |result, state| {
-            if let TxEffect::Skipped(SkippedReason::CannotReserveGas(reason)) = result.tx_receipt {
+            if let TxEffect::Skipped(TxProcessingError::CannotReserveGas(reason)) =
+                result.tx_receipt
+            {
                 assert_eq!(
                     reason,
                     ReserveGasError::<S>::AccountDoesNotExist {
@@ -301,7 +303,9 @@ fn test_reserve_gas_not_enough_balance() {
             })
             .with_max_fee(u64::MAX),
         assert: Box::new(move |result, _state| {
-            if let TxEffect::Skipped(SkippedReason::CannotReserveGas(reason)) = result.tx_receipt {
+            if let TxEffect::Skipped(TxProcessingError::CannotReserveGas(reason)) =
+                result.tx_receipt
+            {
                 assert_eq!(
                     reason,
                     ReserveGasError::<S>::InsufficientBalanceToReserveGas.to_string(),
@@ -345,7 +349,9 @@ fn test_reserve_gas_price_too_high() {
             .with_max_fee(sender_balance)
             .with_gas_limit(Some(GasUnit::from([sender_balance / 2; 2]))),
         assert: Box::new(move |result, _state| {
-            if let TxEffect::Skipped(SkippedReason::CannotReserveGas(reason)) = result.tx_receipt {
+            if let TxEffect::Skipped(TxProcessingError::CannotReserveGas(reason)) =
+                result.tx_receipt
+            {
                 assert_eq!(
                     reason,
                     ReserveGasError::<S>::CurrentGasPriceTooHigh.to_string(),