+++ crates/module-system/sov-modules-stf-blueprint/src/sequencer_mode/unregistered.rs
@@ -5,47 +5,34 @@ use sov_modules_api::capabilities::{
     TransactionAuthorizer, TryReserveGasError, UnregisteredAuthenticationError,
 };
 use sov_modules_api::{
-    DaSpec, ExecutionContext, FullyBakedTx, Gas, GasMeter, GasSpec, PreExecWorkingSet, Spec,
+    DaSpec, ExecutionContext, FullyBakedTx, Gas, GasInfo, GasMeter, PreExecWorkingSet, Spec,
     TxScratchpad, UnlimitedGasMeter, WorkingSet,
 };
 
+use super::registered::{AuthTxOutput, PreExecError};
 use crate::sequencer_mode::common::apply_tx;
-use crate::{ApplyTxResult, Runtime, SkippedReason, TxProcessingError};
+use crate::{ApplyTxResult, Runtime, TxProcessingError};
 
 #[allow(clippy::result_large_err)]
 pub fn process_unauthorized_tx<S: Spec, D: DaSpec, R: Runtime<S, D>>(
     runtime: &R,
-    raw_tx: &FullyBakedTx,
+    auth_output: AuthenticationOutput<
+        S,
+        <R as TransactionAuthenticator<S>>::Decodable,
+        <R as TransactionAuthenticator<S>>::AuthorizationData,
+    >,
+    gas_info: GasInfo<S::Gas>,
     sequencer_da_address: &D::Address,
-    gas_price: &<S::Gas as Gas>::Price,
     height: u64,
-    tx_scratchpad: TxScratchpad<S::Storage>,
+    mut tx_scratchpad: TxScratchpad<S::Storage>,
     execution_context: ExecutionContext,
 ) -> (
     Result<ApplyTxResult<S>, TxProcessingError>,
     TxScratchpad<S::Storage>,
 ) {
-    let mut pre_exec_working_set =
-        tx_scratchpad.to_pre_exec_working_set(UnlimitedGasMeter::new_with_price(gas_price.clone()));
-
-    let (tx, auth_data, message) = match authenticate_unregistered_with_cycle_count(
-        runtime,
-        raw_tx,
-        &mut pre_exec_working_set,
-    ) {
-        Ok(v) => v,
-        Err(e) => {
-            return (
-                Err(TxProcessingError::InvalidUnregisteredTx(e)),
-                pre_exec_working_set.to_scratchpad_and_gas_meter().0,
-            );
-        }
-    };
-
-    let raw_tx_hash = tx.raw_tx_hash;
-    let tx = &tx.authenticated_tx;
-
-    let (mut tx_scratchpad, mut gas_meter) = pre_exec_working_set.to_scratchpad_and_gas_meter();
+    let (auth_tx, auth_data, message) = auth_output;
+    let raw_tx_hash = auth_tx.raw_tx_hash;
+    let tx = &auth_tx.authenticated_tx;
 
     let ctx = match runtime
         .transaction_authorizer()
@@ -54,10 +41,7 @@ pub fn process_unauthorized_tx<S: Spec, D: DaSpec, R: Runtime<S, D>>(
         Ok(ctx) => ctx,
         Err(e) => {
             return (
-                Err(TxProcessingError::Skipped {
-                    reason: SkippedReason::CannotResolveContext(e.to_string()),
-                    raw_tx_hash,
-                }),
+                Err(TxProcessingError::CannotResolveContext(e.to_string())),
                 tx_scratchpad,
             );
         }
@@ -70,47 +54,29 @@ pub fn process_unauthorized_tx<S: Spec, D: DaSpec, R: Runtime<S, D>>(
             .check_uniqueness(&auth_data, &ctx, &mut tx_scratchpad)
     {
         return (
-            Err(TxProcessingError::Skipped {
-                reason: SkippedReason::IncorrectNonce(e.to_string()),
-                raw_tx_hash,
-            }),
+            Err(TxProcessingError::IncorrectNonce(e.to_string())),
             tx_scratchpad,
         );
     }
 
-    if let Err(e) = gas_meter.charge_gas(&S::gas_forced_sequencer_registration_cost()) {
+    if let Err(TryReserveGasError { reason }) = runtime.gas_enforcer().try_reserve_gas(
+        tx,
+        &gas_info.gas_price,
+        ctx.sender(),
+        &mut tx_scratchpad,
+    ) {
         return (
-            Err(TxProcessingError::Skipped {
-                reason: SkippedReason::CannotReserveGas(e.to_string()),
-                raw_tx_hash,
-            }),
-            tx_scratchpad,
-        );
-    }
-
-    if let Err(TryReserveGasError { reason }) =
-        runtime
-            .gas_enforcer()
-            .try_reserve_gas(tx, gas_price, ctx.sender(), &mut tx_scratchpad)
-    {
-        return (
-            Err(TxProcessingError::Skipped {
-                reason: SkippedReason::CannotReserveGas(reason.to_string()),
-                raw_tx_hash,
-            }),
+            Err(TxProcessingError::CannotReserveGas(reason.to_string())),
             tx_scratchpad,
         );
     }
 
     let working_set: WorkingSet<S> =
-        match WorkingSet::try_create_working_set(tx_scratchpad, &gas_meter.gas_info(), tx) {
+        match WorkingSet::try_create_working_set(tx_scratchpad, &gas_info, tx) {
             Ok(ws) => ws,
             Err(err) => {
                 return (
-                    Err(TxProcessingError::Skipped {
-                        reason: SkippedReason::OutOfGas(err.reason.to_string()),
-                        raw_tx_hash,
-                    }),
+                    Err(TxProcessingError::OutOfGas(err.reason.to_string())),
                     err.scratchpad,
                 );
             }
@@ -149,19 +115,35 @@ pub fn process_unauthorized_tx<S: Spec, D: DaSpec, R: Runtime<S, D>>(
     (Ok(apply_tx), tx_scratchpad)
 }
 
+#[allow(clippy::type_complexity)]
+pub(crate) fn authenticate_unregistered_tx<S: Spec, Da: DaSpec, R: Runtime<S, Da>>(
+    runtime: &R,
+    gas_price: &<S::Gas as Gas>::Price,
+    tx: &FullyBakedTx,
+    tx_scratchpad: TxScratchpad<S::Storage>,
+) -> (
+    Result<(AuthTxOutput<S, R>, GasInfo<S::Gas>), PreExecError>,
+    TxScratchpad<S::Storage>,
+) {
+    let mut pre_exec_working_set =
+        tx_scratchpad.to_pre_exec_working_set(UnlimitedGasMeter::new_with_price(gas_price.clone()));
+
+    let res = authenticate_unregistered_with_cycle_count(runtime, tx, &mut pre_exec_working_set);
+    let (tx_scratchpad, gas_meter) = pre_exec_working_set.to_scratchpad_and_gas_meter();
+
+    match res {
+        Err(e) => (Err(PreExecError::UnregisteredAuthError(e)), tx_scratchpad),
+
+        Ok(ok) => (Ok((ok, gas_meter.gas_info())), tx_scratchpad),
+    }
+}
+
 #[cfg_attr(all(target_os = "zkvm", feature = "bench"), cycle_tracker)]
 fn authenticate_unregistered_with_cycle_count<S: Spec, Da: DaSpec, R: Runtime<S, Da>>(
     runtime: &R,
     tx: &FullyBakedTx,
     pre_exec_working_set: &mut PreExecWorkingSet<S, UnlimitedGasMeter<S::Gas>>,
-) -> Result<
-    AuthenticationOutput<
-        S,
-        <R as TransactionAuthenticator<S>>::Decodable,
-        <R as TransactionAuthenticator<S>>::AuthorizationData,
-    >,
-    UnregisteredAuthenticationError,
-> {
+) -> Result<AuthTxOutput<S, R>, UnregisteredAuthenticationError> {
     let auth_input = borsh::from_slice(&tx.data).map_err(|e| {
         UnregisteredAuthenticationError::FatalError(FatalError::DeserializationFailed(
             e.to_string(),